fix(client): resolve Escape kbd shortcut on panels not working
Turns out specifying an array for the shortcut key is interpreted as a serial sequence, not an Or/Union group. And the kbd shortcuts help panel, which was defined with a closeKey of Escape, was causing Escape shortcut commands on other panels to be ignored. Neutering closeKey dynamically didn't work, so I just removed the closeKey shortcut altogether.